refactor: :truck: Deprecate Accordion
Accordion is renamed to LegacyAccordion
