+++ packages/react/src/components/LegacyAccordion/Accordion.module.css
