+++ packages/react/src/components/LegacyAccordion/AccordionContent.tsx
