+++ packages/react/src/components/LegacyAccordion/AccordionHeader.tsx
