+++ packages/react/src/components/LegacyAccordion/AccordionIcon.module.css
