+++ packages/react/src/components/LegacyAccordion/AccordionIcon.tsx
