+++ packages/react/src/components/LegacyAccordion/icons/Arrow.tsx
