+++ packages/react/src/components/LegacyAccordion/icons/CircleArrow.tsx
