+++ packages/react/src/components/LegacyAccordion/icons/index.ts
