+++ packages/react/src/components/LegacyAccordion/index.ts
